Add ability to ignore invalid ids
Added a new keyword argument for get_series which allows for invalid ids
to simply be ignored rather then fail. This will avoid unecessary
requests for series that were valid, but happened to be grouped with a
invalid series.

Had to also rewrite a few of the tests since this new version returns a
dataframe instead of series when one one series is requests.